Se modifica Order y Order Service para que al crear la orden 1) se calculen los descuentos  y 2) se devuelva el detalle del item --> en el detalle devuelto, muestra el precio SIN el descuento, el descuento lo aplica en el subtotal.

+++ src/main/java/com/uade/circulo/entity/Order.java
@@ -39,18 +39,18 @@ public class Order {
     private OrderStatus orderStatus;
 
     @Column(name = "importe_total", nullable = false)
-    private float importeTotal;
+    private double importeTotal;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
     @JsonBackReference
     private User user;
 
-    public float getImporteTotal(){
+    public double getImporteTotal(){
         return importeTotal;
     }
 
-    public void setImporteTotal(float importeTotal){
+    public void setImporteTotal(double importeTotal){
         this.importeTotal = importeTotal;
     }
